now its nearly right

+++ FilthyRich/Country.h
@@ -5,11 +5,11 @@ class Country {
 private:
 	std::string name;
 	int billionares;
-	int sumOfMoney;
+	float sumOfMoney;
 
 public:
 	Country() {}
-	Country(std::string n, int b, int s) : Country() {
+	Country(std::string n, int b, float s) : Country() {
 		this->name = n;
 		this->billionares = b;
 		this->sumOfMoney = s;
@@ -18,11 +18,11 @@ public:
 
 	void setName(std::string n) { this->name = n; }
 	void addBillionare() { this->billionares++; }
-	void setSum(int s) { this->sumOfMoney = s; }
+	void setSum(float s) { this->sumOfMoney = s; }
 
 	std::string getName() { return name; }
 	int getBillionares() { return billionares; }
-	int getSum() { return sumOfMoney; }
+	float getSum() { return sumOfMoney; }
 };
 
 static std::ostream& operator<<(std::ostream &os, Country &c) {
